Tested ADC module (intRef)
This example shows how to use the internal reference of REF module using the REF control registers. The ADC12_A uses the internal 1.5V reference and performs a single conversion on channel A0. The conversion results are stored in ADC12MEM0. Test by applying a voltage to channel A0, then setting and running to a break point at the "__no_operation()" instruction. To view the conversion results, open an ADC12_A register window in debugger and view the contents of ADC12MEM0.